use tabs to switch launch model types between LLM and embeddings
