+++ xinference/web/ui/src/App.js
@@ -7,7 +7,6 @@ import Layout from "./scenes/_layout";
 import RunningModels from "./scenes/running_models";
 import LaunchModel from "./scenes/launch_model";
 import RegisterModel from "./scenes/register_model";
-import LaunchEmbedding from "./scenes/launch_embedding";
 
 function App() {
   const [theme] = useMode();
@@ -22,7 +21,6 @@ function App() {
                 <Route path="/" element={<LaunchModel />} />
                 <Route path="/running_models" element={<RunningModels />} />
                 <Route path="/register_model" element={<RegisterModel />} />
-                <Route path="/launch_embedding" element={<LaunchEmbedding />} />
               </Route>
             </Routes>
           </ApiContextProvider>
